split up action entity classes
