+++ src/qz/printer/action/PrintPDF.java
@@ -18,8 +18,8 @@ import org.codehaus.jettison.json.JSONObject;
 import qz.common.Constants;
 import qz.printer.PrintOptions;
 import qz.printer.PrintOutput;
-import qz.printer.action.entity.BookBundle;
-import qz.printer.action.entity.PDFWrapper;
+import qz.printer.action.pdf.BookBundle;
+import qz.printer.action.pdf.PDFWrapper;
 import qz.utils.ConnectionUtilities;
 import qz.utils.PrintingUtilities;
 import qz.utils.SystemUtilities;
